reader/requirements.txt
aiostream
cryptography
dataclasses_json
docopt
dotmap
flask
json5
jsonref
matplotlib
minimalmodbus
opencv-python
pyjwt[crypto]
pyserial
pyudev
pyusb
pyzmq
python-jose
requests
scipy
ujson
websockets
auth0-python
